Changed long to int fot loop

+++ src/DataTypes.java
@@ -5,7 +5,7 @@ public class DataTypes {
 
         long s = 0;
         // below is a "foreach" loop which iterates through numbers
-        for (long x : numbers) {
+        for (int x : numbers) {
             s += x;
         }
         return s;
